splitting up types into different folders done
