+++ src/features/auth/hooks/useRegister.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
 import { useAuth } from "@/features/auth/context/AuthContext";
+import { useState } from "react";
 
 export const useRegister = () => {
   const [error, setError] = useState<string | null>(null);
