+++ src/features/nav/types.ts
@@ -1,0 +1,12 @@
+export type Notification = {
+  id: number;
+  type: string;
+  content4: string;
+  content: string | undefined;
+  createdAt: string | number | Date;
+  user: string | undefined;
+  read: boolean;
+  _id: number;
+  content2: string;
+  content3: string;
+};
