+++ src/pages/Search/components/SearchLists.tsx
@@ -1,6 +1,7 @@
+import { Media } from "@/pages/MediaPage/types";
 import { useFetch } from "@/hooks/useFetch";
 import { Container } from "@mui/system";
-import { ListToParse, Media } from "@/types/types";
+import { ListToParse } from "@/pages/List/types";
 import { Card, CardContent, CardMedia, Typography } from "@mui/material";
 
 const SearchLists = () => {
