+++ src/pages/Search/types.ts
@@ -1,0 +1,11 @@
+import { Media } from "../MediaPage/types";
+
+export type Result = {
+  data: {
+    results: Media;
+  };
+};
+
+export type query = {
+  query: string | undefined;
+};
